moved requests to helpers file
