+++ lib/helpers.js
@@ -21,6 +21,7 @@ var format = require('util').format;
 var debug = require('debug')('generator-swiftserver:common');
 var fs = require('fs');
 var path = require('path');
+var request = require('request');
 
 // Keywords which are reserved in Swift (taken from the Language Reference)
 var RESERVED_DECLARATION_KEYWORDS = [
@@ -540,4 +541,72 @@ exports.getBluemixDefaultPlan = function(serviceType) {
     case 'pushnotifications':   return 'Lite';
     default:                    return 'Lite';
   }
+}
+
+exports.performSDKGeneration = function(sdkName, sdkType, fileContent, callback) {
+
+  // Call SDK generator
+  var baseURL = "https://mobilesdkgen.ng.bluemix.net/sdkgen/api/generator/"
+  var startGenURL = baseURL + sdkName + "/" + sdkType; // ex. ios_swift or server_swift
+  var authToken = "";
+  
+  request.post({
+    headers: {'Accept' : 'application/json',
+              'Content-Type' : 'application/json',
+              'Authorization' : 'Bearer ' + authToken},
+    url:     startGenURL,
+    body: fileContent
+  }, function(error, response, body){
+    body = JSON.parse(body);
+
+    var job = body['job']
+    var genID = job['id'];
+
+    var sdkReady = false
+    function getStatus() {
+      if (!sdkReady) {
+        request.get({
+          headers: {'Accept' : 'application/json',
+                    'Authorization' : 'Bearer ' + authToken},
+          url: baseURL + genID + "/status"
+        }, function(error, response, body){
+          body = JSON.parse(body);
+
+          if (body['status'] === "FINISHED") {
+            console.log("finished");
+            sdkReady = true
+            getSDK();
+          }
+        })
+      }
+    }
+
+    function getSDK() {
+
+      var mkdirp = require('mkdirp');
+      mkdirp("./SDKs", function (err) {
+          if (err) console.error(err)
+      });
+
+      request.get({
+        headers: {'Accept' : 'application/zip',
+                  'Authorization' : 'Bearer ' + authToken},
+        url: baseURL + genID
+      })
+      .pipe(fs.createWriteStream('SDKs/iOS_SDK.zip'))
+      .on('finish', function () {
+        callback();
+      });
+    }
+
+    // Goes through 10 get requests as needed for a 30 second timeout
+    (function myLoop (i) {          
+       setTimeout(function () {   
+          getStatus()
+          if (!sdkReady && --i) myLoop(i); // decrement i and call myLoop again if i > 0
+       }, 3000)
+    })(10);
+
+  });
+
 }
